Refactor: move api folder path
